+++ packages/react-ui/src/shared/Lightbox/index.ts
@@ -1,2 +1,6 @@
-import Lightbox from './Lightbox';
-export default Lightbox;
+import BaseLightbox, { BaseLightboxProps } from './BaseLightbox';
+import Lightbox, { LightboxProps } from './Lightbox';
+
+export default BaseLightbox;
+export { BaseLightboxProps, Lightbox, LightboxProps };
+